Work on metadata renaming
