+++ snowex_db/batch.py
@@ -11,7 +11,8 @@ from snowex_db import db_session
 from snowex_db.interpretation import get_InSar_flight_comment
 from snowex_db.metadata import (DataHeader, SMPMeasurementLog,
                                 read_InSar_annotation)
-from snowex_db.upload import UploadProfileData, UploadRaster
+from snowex_db.upload.layers import UploadProfileData
+from snowex_db.upload.rasters import UploadRaster
 from snowex_db.utilities import assign_default_kwargs, get_logger
